# frontend/openapi.json
{"openapi":"3.1.0","info":{"title":"Semantic Shapes API","description":"API for visualizing and exploring word embeddings","version":"1.0.0"},"paths":{"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/info":{"get":{"summary":"Get Info","description":"Return metadata about the model (type, dimensions, vocab size)","operationId":"get_info_api_info_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/vector":{"get":{"summary":"Get Vector","description":"Return the raw embedding vector for a given word","operationId":"get_vector_api_vector_get","parameters":[{"name":"word","in":"query","required":true,"schema":{"type":"string","title":"Word"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/similar":{"get":{"summary":"Get Similar","description":"Return top-N most similar words (cosine similarity)","operationId":"get_similar_api_similar_get","parameters":[{"name":"word","in":"query","required":true,"schema":{"type":"string","title":"Word"}},{"name":"n","in":"query","required":false,"schema":{"type":"integer","default":10,"title":"N"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/arithmetic":{"get":{"summary":"Vector Arithmetic","description":"Compute vector arithmetic and return top-N nearest words","operationId":"vector_arithmetic_api_arithmetic_get","parameters":[{"name":"expr","in":"query","required":true,"schema":{"type":"string","title":"Expr"}},{"name":"n","in":"query","required":false,"schema":{"type":"integer","default":5,"title":"N"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/projected":{"get":{"summary":"Get Projected","description":"Return 2D or 3D coordinates of words (via PCA or t-SNE)","operationId":"get_projected_api_projected_get","parameters":[{"name":"words","in":"query","required":true,"schema":{"type":"string","title":"Words"}},{"name":"method","in":"query","required":false,"schema":{"type":"string","default":"pca","title":"Method"}},{"name":"dimensions","in":"query","required":false,"schema":{"type":"integer","default":2,"title":"Dimensions"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/vocab":{"get":{"summary":"Get Vocabulary","description":"Return list of available vocabulary words","operationId":"get_vocabulary_api_vocab_get","parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":10000,"title":"Limit"}},{"name":"starts_with","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Starts With"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}